Rename remove_sources -> unload_code, update_sources -> update_code, deploy_sources -> upload_code
