+++ xawscf/commands/delete_stack.py
@@ -7,7 +7,7 @@ from ..utils.client import get_client
 from ..utils.cloudformation import wait, WaiterError
 from ..utils.logger import log
 from ..pattern.pattern import get_pattern
-from .remove_sources import run as call_remove_sources
+from .unload_code import run as call_unload_code
 
 def delete_stack(pattern):
     stack_name = pattern.get('project')
@@ -18,10 +18,10 @@ def delete_stack(pattern):
     except WaiterError as err:
         raise RuntimeError(err.last_response)
 
-def run(remove_sources=False):
+def run(unload_code=False):
     log('Deleting stack')
     pattern = get_pattern()
     delete_stack(pattern)
     log('Done')
-    if remove_sources:
-        call_remove_sources()
+    if unload_code:
+        call_unload_code()
