+++ xawscf/commands/pack.py
@@ -10,7 +10,7 @@ from ..utils import helper
 from ..utils.logger import log
 from ..utils.parallel import run_parallel
 from ..pattern.pattern import get_pattern
-from .deploy_sources import run as call_deploy_sources
+from .upload_code import run as call_upload_code
 
 def pack_directory(zf, real_dir, zip_dir):
     for item in os.listdir(real_dir):
@@ -44,10 +44,10 @@ def build_packages(pattern, names):
     functions = helper.select_functions(pattern, names)
     run_parallel(map(get_task, helper.get_code_uri_list(functions)))
 
-def run(deploy_sources=False, names=None):
+def run(upload_code=False, names=None):
     log('Packing sources')
     pattern = get_pattern()
     helper.ensure_folder()
     build_packages(pattern, names)
-    if deploy_sources:
-        call_deploy_sources()
+    if upload_code:
+        call_upload_code()
